[sql] Don't force a default query parallelism 2. (#3515)
Currently we set the default query parallelism in datafusion to 2.
This parameter controls the number of parallel work that datafusion
does while processing a single query.
We used to set it to the minimum possible to aggressively limit the
resources used while processing a query, however recent work around
datafusion and rocksdb has affective limits set elsewhere,
and with that the query duration has become noticeably slower.
With this commit, we remove this default and let datafusion to use
the number of cores as the default parallelism.
And we still have the configuration key to limit that.