Restrict FSM table reads to PartitionStore only
This change suggests that we should start splitting operations that require partition-store (read committed) access from operations that operate on the write batch. In fact, I'm hoping that at
some point we'll have a much narrower surface area for write-batch based reads.

The current trait design leads a lot of duplicated code and blurs the line between what needs to operate on an open transaction data vs. operating on the committed data. I'm hoping we can change that over time.