Force loading durable_lsn from persistent storage
This change gives us a reliable way to be sure that we are reading the applied_lsn from the durable storage (flushed to disk) when it's needed. This
fixes the potential issue that can happen if we close and re-open a partition store without flushing.

Note: this doesn't remove the benefit of having the event listener on rocksdb, the event listener allows us to know the latest durable_lsn without
constantly forcing a read from the durable storage. In practice, we only need to read the durable_lsn (the durably stored applied_lsn) once when opening the partition_store.