Adds per-partition trim queue tracking
NOTE: This is current unused but will be used in a future PR when trims are executed by the partition leader.

The design approach:
1. The partition stores the last partition durability point in FSM + timestamp (when it was declared as durable).
2. The partition (will) issue a trim for the log up to the last durable point periodically. (e.g. every 1 minute)

Now, to add support for delayed trims. The approach I'm taking here is to only trim if the durability point is X seconds old. This value is configurable.
The trim queue is a best-effort approach to keep information about the older durability points in the partition, this allows us to still execute older durability before the cutoff point when the latest durability point is not old enough to get trimmed.

The downside of this approach is that we'll have scenarios where we don't know the older durability points (mainly at certain reconfiguration scenarios, and when we are not replaying from the log). In that case the log trimming will lag behind until the latest durable point is old enough to get trimmed.
A pathological case is when this happens in a continuous fashion such that the sliding window of trim queue is never filled and we never trim. This should be a rare scenario but it's theoretically possible.

```
// intentionally empty
```