Force loading durable_lsn from persistent storage on store open
This change gives us a reliable way to be sure that we are reading the applied_lsn from the durable storage (flushed to disk) when it's needed. This
fixes the potential issue that can happen if we close and re-open a partition store without flushing.

Note: this doesn't remove the benefit of having the event listener on rocksdb, the event listener allows us to know the latest durable_lsn without
constantly forcing a read from the durable storage. In practice, we only need to read the durable_lsn (the durably stored applied_lsn) once when opening the partition_store.

Additionally:
- Moves reporting of durable_lsn to the partition processor instead of partition procesosr manager.
- Introduce a watch for durable LSN, the primary use at the moment is to detect if durable_lsn has changed or not before we notify ReplicaSetState. In a possible future, the watch can be used to directly update the ReplicaSetState without relying on the status update timer, although the benefits of doing this are unclear.
- Access to get_durable_lsn is now through PartitionProcessor.

```
// intentionally empty
```