[experimental feature] Partition-driven log-trimming
This changeset introduces the experimental feature `worker.experimental-partition-driven-log-trimming` that disables cluster-controller-driven log trimming and replaces it with
a new component (DurabilityTracker) that runs on the partition leader. The durability tracker respects the configured `worker.durability_mode` (hidden in config still).

DurabilityTracker in conjunction with DurabilityMode allow partitions to self-report their "durable" lsn via a WAL message `UpdatePartitionDurability`. Once the durability point is observed
by the leader, it gets added to the trim-queue which tracks when this point was observed as durable. The trim-queue is driven by the leader to time-shift the actual trim operation based on
`worker.trim-delay-interval`. This allows fine-grained control over (a) when do we consider a partition to be fully "durable" and (b) when do we trim the log when this happens.


The feature is disabled by default as it depends on `UpdatePartitionDurability` becoming rollback-safe.

To test, set `worker.experimental-partition-driven-log-trimming = true` in config of all nodes, and optionally set `worker.trim-delay-interval = "5m"` to delay trimming by 5 minutes.

```
// intentionally left empty
```