Correct Scheduler::requires_reconfiguration checks
This commit corrects the Scheduler::requires_reconfiguration checks which
previously could cause false positives. The problem was that after checking
that next does not require changes, we checked current but did not compare the
proposed reconfiguration to a possibly set next value. Now we only check current
if there is no reconfiguration ongoing (no next partition configuration is set).