Complete reconfiguration if all current workers are disabled
This commit extends when to complete a partition reconfiguration by checking
whether all current workers are disabled. If this is the case (e.g. when all
current nodes have been removed from the cluster), we can complete the
reconfiguration because we are sure that none of the current nodes will come
back. Hence, there is no need to wait for any of the next pps to become active.

This additional condition helps us fix a problem where pps reading from a partially
sealed loglet get stuck if there is nobody writing to the log and the system requires
any of the catching up pps to become the leader (because all current workers are
permanently removed from the cluster).

This fixes #3529.