[fix] Fixes lookup index management of replicated loglet
This fixes a couple of bugs in lookup index management in the log chain, the only impact in production code is that the index for a loglet might remain present after the loglet has been trimmed, the other issues do not impact current production code.


Test Plan:
Added unit tests to cover the lookup index management code.