Update locally stored NodesConfiguration when leaving the metadata store cluster
The Raft metadata server stores a local copy of the NodesConfiguration when it learns about
being a member of the cluster because we retrieve the peer addresses from the NodesConfiguration
and want to remember this information on restarts. Due to this, it can happen after a node is removed
and being restarted, that it sees an old version of the NodesConfiguration where it is still
MetadataServerState::Member. As a consequence the Raft metadata server might try to rejoin the cluster
even though it was removed before. We solve this problem by updating the local copy of the NodesConfiguration
when leaving the metadata store cluster (becoming a MetadataServerState::Standby).

This fixes #3535.