Check that selected node sets fulfill the desired replication properties
This commit adds a check that verifies that a selected node set fulfills the
desired replication property before reconfiguring a log. This ensures that we
don't end up in a situation where a log becomes unwritable because there is
no chance to find a write quorum in the node set.

This fixes #3564.