Ensure that RocksDbManager::close_db_tasks is used from within a Tokio context when spawning
RocksDbManger::background_close_db spawns blocking tasks on a TaskTracker. For this to work, it
needs to run from within a Tokio context. Since we call backgrond_close_db when RocksAccess is
dropped, we need to handle the situation where this happens from outside of a Tokio context.
Therefore, this commit tries to establish the Tokio context by calling TaskCenter Handle::block_on
which is runs the given future on the default Tokio runtime.

This fixes #3573.