[Bifrost][experimenta] Add support for chain sealing marker
This changeset adds support for chain sealing marker. An experimental feature (to be enabled in v1.6) that allows a log chain to be sealed with a special marker in metadata without requiring a new segment (reconfiguration).

The sealing marker enables readers to detect that a log chain is partially sealed and in future PRs it will be able to seal the chain to allow readers to reach the tail of the log without requiring a reconfiguration to take place.
This also will allow the creation of empty logs (log chains with no segments), or logs to trim all their segments which can be useful in tricky snapshot+reconfiguration scenarios where historical nodesets would cause safety checker
to trip even that all data has been trimmed if this is the last segment.

Additionally, this will allow the creation of a new cluster that's seeded with initial LSN offsets to allow seamless restore of partition stores without requiring historical fake segments.

To facilitate this, a new `SealedLoglet` kind has been added but it's hidden from all user-facing APIs/surfaces by duplicating the ProviderKind enum into two. A `ProviderKind` is still used to for user-facing surfaces,
and `InternalKind` is the union of ProviderKind and special kinds (currently only `Sealed`).


The sealed segment has a special property that it shares the same segment index as the previous segment.

In this changeset, only `bifrost.find_tail()` will automatically seal the chain if the underlying loglet is sealed while the chain is still open (gated by the experimental flag). In future PRs, this will also happen by readers when they detect
partial sealing (reconfiguration) that is seemingly stuck for extended time.

Also a new `restatectl` command `restatectl log seal <LOG-IDs>` has been added to trigger chain sealing from the CLI (and a corresponding ClusterCtrl gRPC API).


```
// intentionally empty
```