Fix bug in Scheduler::complete_reconfiguration which compared the wrong version
By comparing the next version of the current epoch metadata with itself, we were always completing
a reconfiguration. This could have led to completing a reconfiguration prematurely (while it did not
fulfill the should_complete_reconfiguration criteria). Additionally, this commit improves the logging
when transitioning from one partition configuration to another.