[observability] Disable histogram idle timeout by default
In recent restate release we changed the default idle timeout of histograms to 3 minutes, this caused some of the histograms to drop after being idle and never appearing again due to this issue in metrics-rs https://github.com/metrics-rs/metrics/issues/372.

The issue happens for histograms that we keep a stable handle for, we keep those handles to avoid the repetitive heap allocation in our hot paths. Although we are disabling the default idle timeout, users will still see the issue if they are setting this value in configuration files.

This should be considered as a short-term mitigation, we'll need a longer-term solution still.

```
// intentionally empty
```