[WIP] PartitionStoreManager exclusive ownership by Worker role
This allows the worker to flush rocksdb as soon as the worker role is stopped and before the rest of the system is shut down. In order to achieve this, datafusion queries will always use the remote scanner even on worker nodes. This adds serialization/memory cost that would have been otherwise avoided but potentially can be optimized in future PRs.