[fingerprint] Automatically generate cluster fingerprint for existing clusters
This changeset allows nodes starting with v1.5.0 to auto provision a cluster fingerprint for their cluster. Once a fingerprint is generated, it's immutable and cannot be changed by other nodes.

The cluster fingerprint is exchanged through the node's configuration propagation and nodes will eventually converge into using it for network connection checking.

Note that this doesn't cause any network connectivity issues during the migration because networkings only enforces the fingerprint check if the node is aware of its fingerprint **and** the incoming connection carries the fingerprint in the Hello message.

This has been manually tested and verified to work with rolling upgrades/downgrades and with full downgrade to 1.4.x without losing the generated fingerprint.