[fingerprint] Reject metadata updates that change the cluster fingerprint
This adds validation to the metadata manager to ensure that a node that knows about its cluster fingerprint will reject any nodes configuration that unsets or mutates the fingerprint. This sanity check
ensures that a shared metadata store that has been corrupted by a manual change (or an overwrite from another cluster) will not poison the state of the node.