[fingerprint] Validate cluster fingerprint of snapshots
This changeset allows nodes to validate snapshot's metadata against the cluster's fingerprint if the snapshot **and** the node both have a fingerprint set. This is designed this way to allow nodes to continue
to read older snapshots although they don't have the fingerprint set. New snapshots will carry the cluster snapshot.

It also fixes a few issues where with error logging when this case happens, making it easier to debug. A screenshot of the error is attached in a comment.