Extension trait for `Duration` for friendly display
Introduces a crate for time utilities (`serde-time-util`). The crate currently provides a trait `DurationExt` which provides a `friendly()` method for formatting a `Duration` as a human-readable string. It also allows parsing a human-friendly duration into a string.


Additionally, this removes the helpers from `DurationString` and keeps `restate-serde-util`'s DurationString as a serialization helper only.

Minor but relevant changes:
- Some debug traces that reported the response latency now use the `to_seconds_span()` since it's cheaper to compute.
- Fixed typos in restate CLI service config commands
- Migrateed all uses of `DurationString` to `DurationExt::friendly()` including in CLI argument parsing.