[PartitionStore] Decouple read and write invocation status traits
This is a step towards replacing all the table traits with individual read/write traits. The separation
enables us to be selective about the API shape available when operating on a transaction/write-batch vs. on
the partition store itself. In particular, we would remove the need for `async` when applying changes to the
write-batch. This makes it easier to spot where we might be performing blocking operations. For instance,
if a write is implemented directly on partition store, then it would need a different API/trait to reflect that
it is async-blocking.