Deprecate localhost:9070/cluster-health endpoint
The /cluster-health endpoint should not have been publicly exposed. It is no longer
used by Restate. Hence, this commit deprecates this endpoint by removing it from the
OpenAPI spec as a first step. In v1.7.0, we can then remove this endpoint. The release
notes of v1.6.0 should mention the deprecation of this endpoint so that users can
stop using it.

This fixes #3898.