Introduce version barrier for enabling journal v2 by default
The version barrier requiring min Restate version v1.6.0 will enable the
journal v2 by default. It will be written when becoming leader and running
at least v1.7.0.

Writing the journal v2 by default changes the logic of the state machine. That's
why we are guarding it with a version barrier so that during a rolling upgrade the
state machine state won't diverge.

A concrete problem is that Restart as new with a non-zero prefix would fail on a
node that didn't store the journal entries in the journal v2 table.

Moreover, when rolling back to the previous version (where we are reading from
journal v1 first), there is the situation in which the invocation task would hang.
The problem is that the InvokerStorageReaderTransaction wouldn't read the journal
entries of an unpinned invocation that has an associated journal v2  because it
would only try to read the journal v1.