Introduce kafka-oidc feature and enable it by default for restate-server builds
To use the Kafka ingress with OIDC authentication, we need to enable the gssapi-vendored
and curl-static features on our rdkafka dependency. To make things properly work with cross
compilation, we need to patch rust-sasl and rdkafka and bump librdkafka to v2.12.1. The latter
allows us to compile with curl-static which saves us the hassle to dynamically link against
libraries from a different target architecture.

Update transitive dependencies to work with cross compilation

Bump rdkafka to use librdkafka 2.12.1