Transfer leadership if shutting down as the leading member
If the leading member is shutting down, we should try transferring the leadership
to an alive node in order to speed up the leadership change in the metadata cluster.
Otherwise, we have to wait for the leader election timeout to occur. Note, this is
an optimization that might not work if the node shuts down w/o sending the transfer
leadership message to the target node.

This fixes #3956.