Remove last remnants of local metadata store from the code base
This commit removes the remaining bits of the local metadata stoer from the
code base. This includes the migration logic contained in the replicated
metadata store. The reason this is ok is because migration should have run
in v1.4 or v1.5 and we expect users to never jump minor versions when upgrading.
We keep a minimal safety check which panics if it detects that a migration
might be needed but wasn't run.

This fixes #4040.