[5/N][VQueues] Prototype of integrating invoke/end paths
The new invoke path is gated behind `experimental-enable-vqueues` config flag and it's disabled (and hidden) by default.


As agreed, this PR is _not_ meant to be the definitive implementation of the new invoke path but rather the first step to get the basic call/response path working. This doesn't handle scheduled invocations correctly yet, nor any other paths (suspend, resume, kill, etc.).

The vqueue id and parents are hardcoded for now, this will be changed in a follow-up PR.