Add support for state mutations when using vqueues
This commit integrates vqueues with external state mutations. The way it works is by
storing the external state mutation in the VQueueItems table until it gets run
by the scheduler. When running the state mutation, we retrieve it from the VQueueItems
table and apply the state mutation changes.

The external state mutation EntryId is currently derived from the internal inbox_seq_number
to generate unique EntryId's. In the future we might give external state mutations an explicit
id which could be used to derive the EntryId.

Currently, the DRR scheduler reserves a concurrency token for running state mutations.
We might want to relax this condition if we see that applying state mutations does not
need to be protected by the global concurrency limit.