Remove invocation epoch
This commit removes the invocation epoch from the WAL command messages, the InFlightInvocationMetadata
and the invoker.

It is safe to remove the invocation epoch and the completion_range_epoch_map from the
InFlightInvocationMetadata because the Protobuf storage type will fill in default values when v1.5 reads
v1.6 data. When v1.6 reads v1.5 data, it will simply ignore those fields.

It is safe to remove the invocation epoch from the WAL command types because the field has a #[serde(default)]
attribute in v1.5. Since we are using flexbuffers and bilrost for serde, v1.6 can ignore the invocation epoch
fields from earlier versions.