Improve the error message when cluster fingerprint is not set
This commit improves the panic message when the cluster fingerprint is not set
and it is accessed. If the NodesConfiguration is valid, then this indicates
that the user is resuming the Restate node from a version that didn't write the
cluster fingerprint in the NodesConfiguration (< v1.5). If this is detected, we
are generating a more expressive error message for our users.