[18/N][VQueues] VisibleAt can use MillisSinceEpoch directly
VQueues can now leverage the niche optimization of MillisSinceEpoch and use it directly in places where
the UniqueTimestamp was used in places where happens-before relationship is not required. Previously,
UniqueTimestamp was used as a workaround to efficiently store the physical clock values in Option<> without
incurring the overhead of the Option<> itself. This is now no longer required.